refactored for parent catalog updates
