+++ lambda_functions/stac/stac.py
@@ -82,11 +82,12 @@ def stac_handler(event, context):
         obj = s3.Object(bucket, stac_s3_key)
         obj.put(Body=json.dumps(stac_item))
 
-        # Update parent catalogs
-        update_parent_catalogs(stac_s3_key, s3, bucket)
-
 
 def get_bucket_and_key(message):
+    """
+    Parse the bucket and s3 key from the SQS message
+    """
+
     s3_event = json.loads(message["body"])["Records"][0]
     return s3_event["s3"]["bucket"]["name"], s3_event["s3"]["object"]["key"]
 
@@ -165,110 +166,10 @@ def valid_coord_to_geojson(valid_coord):
 
 
 def get_stac_item_parent(s3_key):
+    """
+    Parse the parent stac catalog from the given s3 key
+    """
     template = '{prefix}/x_{x}/y_{y}/{}'
     params = pparse(template, s3_key).__dict__['named']
     key_parent_catalog = f'{params["prefix"]}/x_{params["x"]}/y_{params["y"]}/catalog.json'
     return f'{GLOBAL_CONFIG["aws-domain"]}/{key_parent_catalog}'
-
-
-def update_parent_catalogs(s3_key, s3_resource, bucket):
-    """
-    Assumed structure:
-        root catalog
-            -> per product/catalog.json
-                -> x/catalog.json
-                    -> y/catalog.json
-    """
-
-    template = '{prefix}/x_{x}/y_{y}/{}'
-    params = pparse(template, s3_key).__dict__['named']
-    y_catalog_name = f'{params["prefix"]}/x_{params["x"]}/y_{params["y"]}/catalog.json'
-    y_obj = s3_resource.Object(bucket, y_catalog_name)
-
-    try:
-        # load y catalog dict
-        y_catalog = json.loads(y_obj.get()['Body'].read().decode('utf-8'))
-
-    except s3_resource.meta.client.exceptions.NoSuchKey as e:
-
-        # The object does not exist.
-        y_catalog = create_y_catalog(params["prefix"], params["x"], params["y"])
-
-        # Potentially x catalog may not exist
-        update_x_catalog(s3_key, s3_resource, bucket)
-
-    # Create item link
-    item = {'href': f'{GLOBAL_CONFIG["aws-domain"]}/{s3_key}',
-            'rel': 'item'}
-
-    # Add item to catalog
-    y_catalog["links"].append(item)
-
-    # Put y_catalog dict to s3
-    obj = s3_resource.Object(bucket, y_catalog_name)
-    obj.put(Body=json.dumps(y_catalog))
-
-
-def create_y_catalog(prefix, x, y):
-    y_catalog_name = f'{prefix}/x_{x}/y_{y}/catalog.json'
-    x_catalog_name = f'{prefix}/x_{x}/catalog.json'
-    return OrderedDict([
-        ('name', f'{prefix}/x_{x}/y_{y}'),
-        ('description', 'List of items'),
-        ('links', [
-            {'href': f'{GLOBAL_CONFIG["aws-domain"]}/{y_catalog_name}',
-             'ref': 'self'},
-            {'href': f'{GLOBAL_CONFIG["aws-domain"]}/{x_catalog_name}',
-             'rel': 'parent'},
-            {'href': GLOBAL_CONFIG["root-catalog"],
-             'rel': 'root'}
-        ])
-        ])
-
-
-def update_x_catalog(s3_key, s3_resource, bucket):
-    template = '{prefix}/x_{x}/y_{y}/{}'
-    params = pparse(template, s3_key).__dict__['named']
-    y_catalog_name_abs = f'{params["prefix"]}/x_{params["x"]}/y_{params["y"]}/catalog.json'
-    y_catalog_name_rel_to_x = f'y_{params["y"]}/catalog.json'
-    x_catalog_name = f'{params["prefix"]}/x_{params["x"]}/catalog.json'
-    x_obj = s3_resource.Object(bucket, x_catalog_name)
-
-    try:
-
-        # load x catalog dict
-        x_catalog = json.load(x_obj.get()['Body'].read().decode('utf-8'))
-
-    except s3_resource.meta.client.exceptions.NoSuchKey as e:
-
-        # The object does not exist.
-        x_catalog = create_x_catalog(params["prefix"], params["x"])
-
-    # search y catalog link
-    for link in x_catalog["links"]:
-        if link["href"] in (y_catalog_name_abs, y_catalog_name_rel_to_x):
-            return
-
-    # y catalog link not found so update it
-    x_catalog["links"].append({"href": y_catalog_name_abs, "rel": "child"})
-
-    # Write back x catalog
-    x_obj.put(Body=json.dumps(x_catalog))
-
-
-def create_x_catalog(prefix, x):
-    # ToDo: check product catalog name/parent to x
-
-    x_catalog_name = f'{prefix}/x_{x}/catalog.json'
-    return OrderedDict([
-        ('name', f'{prefix}/x_{x}'),
-        ('description', 'List of Sub Directories'),
-        ('links', [
-            {'href': f'{GLOBAL_CONFIG["aws-domain"]}/{x_catalog_name}',
-             'ref': 'self'},
-            {'href': f'{GLOBAL_CONFIG["aws-domain"]}/{prefix}/catalog.json',
-             'rel': 'parent'},
-            {'href': GLOBAL_CONFIG["root-catalog"],
-             'rel': 'root'}
-        ])
-        ])
